+++ changelog.md
@@ -29,6 +29,8 @@ Codion Change Log
 - EntityConnection.queryCache() renamed cacheQueries().
 ### is.codion.framework.db.local
 - DefaultLocalEntityConnection, DefaultLocalEntityConnectionProvider, minor logging improvements.
+### is.codion.framework.model
+- AbstractEntityEditModel, DefaultEntitySearchModel, minor logging improvements.
 ### is.codion.swing.framework.ui
 - EntityTableExportModel.all() added, EntityTableExportPanel all/selected radio button setup improved.
 ### is.codion.plugin.flatlaf.intellij